requirements.txt
requests>=2.31.0
pyyaml>=6.0
python-docx>=1.1.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
